% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austres_df.R
\docType{data}
\name{austres_df}
\alias{austres_df}
\title{Quarterly Time Series of the Number of Australian Residents.}
\format{
A tibble with 89 rows and 3 variables:
\describe{
\item{time}{dbl Year and its quarter.}
\item{value}{dbl Australian population value in thousands.}
\item{res_mil}{dbl Australian population value in millions.}
}
}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/00Index.html}
}
\usage{
austres_df
}
\description{
Numbers (in millions) of Australian residents measured quarterly from March 1971 to March 1994. The object is of class "ts".
}
\keyword{datasets}
