% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedsR-package.R
\docType{package}
\name{speedsR-package}
\alias{speedsR}
\alias{speedsR-package}
\title{speedsR: Data Package With Sports-Specific Datasets For The AIS 'SPEEDS' Project}
\description{
Provides a collection of sports-specific datasets as part of the Australian Institute of Sport 'SPEEDS' project. The 'speedsR' package facilitates easy access to these datasets, enabling analysis and research within a sports science context. Available for download from the AIS 'SPEEDS' GitHub repository, it integrates seamlessly as a library in any R-based IDE.
}
\author{
\strong{Maintainer}: Sports Data Analytics Team La Trobe University \email{ais.speeds@gmail.com} (\href{https://orcid.org/0000-0001-6760-6670}{ORCID})

}
\keyword{internal}
